move enum base support to base-application.
