+++ generators/base/support/index.mts
@@ -20,6 +20,5 @@ export { textToArray, stringNullOrEmpty, isSimpleText, htmlEncode, stripMargin, 
 export { parseCreationTimestamp, resetFakerSeed } from './configuration-helpers/sequences.mjs';
 export { default as httpsGet } from './remote.mjs';
 export { isReproducible } from './configuration-helpers/options.mjs';
-export { default as getEnumInfo } from './configuration-helpers/enum.mjs';
 export { removeFieldsWithUnsetValues } from './config.mjs';
 export * from './write-files.mjs';
