+++ generators/client/entity-files.mjs
@@ -17,7 +17,7 @@
  * limitations under the License.
  */
 import { CLIENT_MAIN_SRC_DIR } from '../generator-constants.mjs';
-import { getEnumInfo } from '../base/support/index.mjs';
+import { getEnumInfo } from '../base-application/support/index.mjs';
 
 const CLIENT_COMMON_TEMPLATES_DIR = 'entity/common';
